+++ NEWS.md
@@ -26,3 +26,4 @@
 
 ## 1.0.1
 * Add importFrom progress progress_bar to NAMESPACE to address note in checks
+* Add option to plotting function to remove noise edges if JANE was run with noise_weights = TRUE